Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ char* welcome() {
 
 int main() {
     // printf("%s", welcome());
-    int a,b,c;
+    float a,b,c;
     scanf("%f %f %f",&a,&b,&c);
     float average = (a+b+c)/3;
     printf("Average: %.2f",average);
